config validation first pass
